Created cron container
Cron was not staring alongside postgres. "CMD cron" stopped postgres from starting, and vice versa.

+++ cleanup.sh
@@ -4,5 +4,4 @@
 export PGPASSWORD=password
 
 # Connect to the PostgreSQL database and execute the cleanup SQL query
-psql -h postgres-frcm_postgres_1 -U admin -d fireriskdb \
--c "DELETE FROM fireriskdata WHERE CURRENT_TIMESTAMP >= ExpirationDate;"
+psql -h postgres-frcm_postgres_1 -U admin -d fireriskdb -c "DELETE FROM fireriskdata WHERE CURRENT_TIMESTAMP >= ExpirationDate;"
